requirements.txt
chardet==2.3.0
click==6.6
Django==2.1.4
django-ipware==2.1.0
gpiozero==1.4.1
idna==2.2
ipaddress==1.0.17
oauthlib==2.0.1
picamera==1.13
picraft==1.0
piglow==1.2.4
pigpio==1.38
Pillow==4.0.0
PyQRCode==1.2.1
pyserial==3.2.1
pytz==2018.7
qrcode==6.0
requests==2.12.4
requests-oauthlib==0.7.0
RPi.GPIO==0.6.5
simplejson==3.10.0
six==1.12.0
urllib3==1.19.1
